Relocate and use path aliases js_only
