+++ assets/src/domain/eventEditor/ui/tickets/hooks/useTicketRegistrationCount.ts
@@ -1,4 +1,4 @@
-import { EntityId } from '../../../services/apollo/types';
+import { EntityId } from '@edtrServices/apollo/types';
 
 const useTicketRegistrationCount = (ticketId: EntityId): number => {
 	// @todo get registration count
