--- assets/src/domain/eventEditor/ui/tickets/ticketCard/styles.ts
@@ -1,12 +1,0 @@
-import { CSSProperties } from 'react';
-
-export const btnsStyle: CSSProperties = {
-	bottom: '.5rem',
-	position: 'absolute' as 'absolute',
-	right: '.5rem',
-	textAlign: 'right',
-};
-
-export const priceStyle: CSSProperties = {
-	color: 'grey',
-};
