+++ assets/src/domain/eventEditor/ui/tickets/ticketsList/AddNewTicketButton.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { TagsOutlined } from '@ant-design/icons';
 import { __ } from '@wordpress/i18n';
 
-import { EspressoButton } from '../../../../../application/ui/input';
-import { useEditorModal } from '../../../../../application/ui/layout/editorModal';
+import { EspressoButton } from '@appInputs/index';
+import { useEditorModal } from '@appLayout/editorModal';
 
 const AddNewTicketButton: React.FC = () => {
 	const { openEditor } = useEditorModal();
