+++ assets/src/domain/eventEditor/ui/tickets/ticketsList/actionsMenu/EditTicketButton.tsx
@@ -3,9 +3,9 @@ import { EditOutlined } from '@ant-design/icons';
 import { __ } from '@wordpress/i18n';
 
 import { EspressoButton } from '@application/ui/input';
-import { EditItemButtonProps } from '../../../interfaces/types';
-import { useTicketContext } from '../../../hooks';
-import { useEditorModal } from '../../../../../application/ui/layout/editorModal';
+import { EditItemButtonProps } from '@edtrInterfaces/types';
+import { useTicketContext } from '@edtrHooks/index';
+import { useEditorModal } from '@appLayout/editorModal';
 
 const EditTicketButton: React.FC<EditItemButtonProps> = (props) => {
 	const { id: entityId } = useTicketContext();
