+++ assets/src/domain/eventEditor/ui/tickets/ticketsList/actionsMenu/TicketActionsMenu.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
 import { ActionsMenuComponentProps } from '@appLayout/entityActionsMenu';
-import useTicketsActionMenuItems from '../hooks/useTicketsActionMenuItems';
-import { Ticket } from '../../../services/apollo/types';
+import useTicketsActionMenuItems from '../../hooks/useTicketsActionMenuItems';
+import { Ticket } from '@edtrServices/apollo/types';
 
 const TicketActionsMenu: React.FC<ActionsMenuComponentProps<Ticket>> = ({ entity, menuItemProps, ...menuProps }) => {
 	const menuItems = useTicketsActionMenuItems(entity, menuItemProps);
