+++ assets/src/domain/eventEditor/ui/tickets/ticketsList/cardView/TicketDetails.tsx
@@ -1,13 +1,13 @@
 import React, { useMemo } from 'react';
 import { __ } from '@wordpress/i18n';
 
-import TicketRegistrationsLink from '../TicketRegistrationsLink';
-import useTicketRegistrationCount from '../hooks/useTicketRegistrationCount';
+import TicketRegistrationsLink from '../../TicketRegistrationsLink';
+import useTicketRegistrationCount from '../../hooks/useTicketRegistrationCount';
 import InfinitySymbol from '@application/valueObjects/InfinitySymbol';
 import parseInfinity from '@appServices/utilities/parseInfinity';
 import { EntityDetailsPanel } from '@appDisplay/enityDetails';
 import { MutatorFn } from '@appServices/apollo/mutations';
-import { Ticket } from '../../../services/apollo/types';
+import { Ticket } from '@edtrServices/apollo/types';
 
 interface TicketDetailsProps {
 	ticket: Ticket;
